src/main/java/ru/job4j/loop/Counter.java
package ru.job4j.loop;

public class Counter {
    public static int sum(int start, int finish) {
        int sum = 0;
        for (int i = start; i <= finish; i++) {
            sum += i;
        }
        return sum;
    }

    public static int sumByEven(int start, int finish) {
        int sum = 0;
        for (int i = start; i <= finish; i++) {
            if (i % 2 == 0) {
                sum += i;
            }
        }
        return sum;
    }

    public static void main(String[] args) {
        System.out.println(3);
        System.out.println(2);
        System.out.println(1);
        System.out.println(sumByEven(1, 8));
        System.out.println(sum(3, 8));
    }
}
